Remove persistent CD device udev rules
Persistent CD device names are a not generally usefull on live systems
because many applications expect to find a CD-ROM device as /dev/cdrom.
But this link is only available when booting on the machine where the
system was first booted. On all other machines /dev/cdrom won't be
available. The CD-ROM device will be named /dev/cdromX.

Without this script most media players have to be reconfigured each time
you boot on a different machine.